created and commented out profilePhoto store file and routes - will complete later, and moved About You and Other Links sections to their own files, SRP
