+++ app/api/story_routes.py
@@ -26,7 +26,6 @@ def stories():
 @story_routes.route('/<int:id>')
 def story(id):
     story = Story.query.get(id)
-    print("THIS IS THE STORY ROUTE", story)
     return story.to_dict()
 
 @story_routes.route('/new-story', methods=['POST'])
